Ajout sélection de l'unité

+++ include/listes.h
@@ -14,6 +14,6 @@ void oter_elt(int);
 void ajout_droit(int,vector v);
 void ajout_gauche(int,vector v);
 
-void addUnit(int noPlayer, vector coordUnit);
-void destroyUnit(int noPlayer, vector coordUnit);
-void printList(int noPlayer);
+void addUnit(short noPlayer, vector coordUnit);
+void destroyUnit(short noPlayer, vector coordUnit);
+void printList(short noPlayer);
