+++ include/manageString.h
@@ -2,6 +2,6 @@ void getCoordS(char coordString[], vector * coordUnit);
 char* get2Char(char name[]);
 char* getNameUnit(unitName unit);
 void printNameUnit(unitName unit);
-bool correctCoord(char coordString[], int noPlayer);
+bool correctCoord(char coordString[], short noPlayer);
 int read(char * string, short length);
 void clearBuffer();
