+++ src/units/barrierTotem.c
@@ -4,7 +4,7 @@
 /**
 * Initialise l'unité
 */
-void barrierTotemInit(int noPlayer, vector coordUnit){
+void barrierTotemInit(short noPlayer, vector coordUnit){
 	grid[coordUnit.x][coordUnit.y].stat.HP = 32;
 	grid[coordUnit.x][coordUnit.y].stat.POWER = 0;
 	grid[coordUnit.x][coordUnit.y].stat.ARMOR = 0.0;
