+++ src/units/darkWitch.c
@@ -4,7 +4,7 @@
 /**
 * Initialise l'unité
 */
-void darkWitchInit(int noPlayer, vector coordUnit){
+void darkWitchInit(short noPlayer, vector coordUnit){
 	grid[coordUnit.x][coordUnit.y].stat.HP = 28;
 	grid[coordUnit.x][coordUnit.y].stat.POWER = 24;
 	grid[coordUnit.x][coordUnit.y].stat.ARMOR = 0.0;
