+++ src/units/furgon.c
@@ -4,7 +4,7 @@
 /**
 * Initialise l'unité
 */
-void furgonInit(int noPlayer, vector coordUnit){
+void furgonInit(short noPlayer, vector coordUnit){
 	grid[coordUnit.x][coordUnit.y].stat.HP = 48;
 	grid[coordUnit.x][coordUnit.y].stat.POWER = 0;
 	grid[coordUnit.x][coordUnit.y].stat.ARMOR = 0.0;
