+++ src/units/mudGolem.c
@@ -4,7 +4,7 @@
 /**
 * Initialise l'unité
 */
-void mudGolemInit(int noPlayer, vector coordUnit){
+void mudGolemInit(short noPlayer, vector coordUnit){
 	grid[coordUnit.x][coordUnit.y].stat.HP = 60;
 	grid[coordUnit.x][coordUnit.y].stat.POWER = 20;
 	grid[coordUnit.x][coordUnit.y].stat.ARMOR = 0.0;
